refactor: split db helpers
