+++ scripts/index-lesson-summaries.test.ts
@@ -9,10 +9,10 @@ import { indexLessonSummaries } from './index-lesson-summaries';
 jest.mock('node:fs/promises');
 jest.mock('~lib/turso');
 
-jest.mock<typeof import('~api/notes')>('~api/notes', () => ({
+jest.mock<typeof import('~api/notes-db')>('~api/notes-db', () => ({
+    getNoteMetadata: jest.fn(),
     getAllNotes: jest.fn(),
-    fetchNotes: jest.fn(),
-    getNoteData: jest.fn(),
+    searchNotes: jest.fn(),
 }));
 
 jest.mock<typeof import('./helpers')>('./helpers', () => ({
