+++ scripts/update-md-timestamps.test.ts
@@ -3,10 +3,10 @@ import { updateMdTimestamps } from './update-md-timestamps';
 
 jest.mock('./helpers');
 
-jest.mock<typeof import('~api/notes')>('~api/notes', () => ({
-    fetchNotes: jest.fn(),
+jest.mock<typeof import('~api/notes-db')>('~api/notes-db', () => ({
+    getNoteMetadata: jest.fn(),
     getAllNotes: jest.fn(),
-    getNoteData: jest.fn(),
+    searchNotes: jest.fn(),
 }));
 
 describe('updateMdTimestamps', () => {
